pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spectral-mcp"
version = "0.1.4"
description = "Automatically discover and document web application APIs"
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["mcp", "api", "reverse-engineering", "openapi", "graphql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "requests>=2.28",
    "mitmproxy>=11.0.2",
    "graphql-core>=3.2",
    "compact-json>=1.8.2",
    "mcp>=1.0",
    "jq>=1.11.0",
    "jinja2>=3.1",
    "pydantic-ai-slim[anthropic]>=1.67.0",
]

[project.scripts]
spectral = "cli.main:cli"

[tool.setuptools.packages.find]
include = ["cli*"]

[tool.setuptools.package-data]
"cli" = ["prompts/*.j2", "completions/*"]

[project.optional-dependencies]
android = [
    "mitmproxy>=11.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mkdocs<2",
    "mkdocs-material>=9.7.2",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
]

[tool.ruff.lint]
select = ["I"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true

[project.urls]
Homepage = "https://github.com/romain-gilliotte/spectral"
Documentation = "https://romain-gilliotte.github.io/spectral/"
Repository = "https://github.com/romain-gilliotte/spectral"
Issues = "https://github.com/romain-gilliotte/spectral/issues"

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
commit_message = "chore(release): v{version}"
